Create common base class for Kismet Variable expressions
